restructure preprocessing folder
