+++ resurfemg/helper_functions.py
@@ -78,232 +78,232 @@ class Range(namedtuple('RangeBase', 'start,end')):
         return slice(*map(int, self))   # maps whole tuple set
 
 
-def emg_bandpass_butter(data_emg, low_pass, high_pass):
-    """The parameter taken in here is the Poly5 file. Output is
-    the EMG after a bandpass as made here.
+# def emg_bandpass_butter(data_emg, low_pass, high_pass):
+#     """The parameter taken in here is the Poly5 file. Output is
+#     the EMG after a bandpass as made here.
 
-    :param data_emg: Poly5 file with the samples to work over
-    :type data_emg: ~TMSiSDK.file_readers.Poly5Reader
-    :param low_pass: The number to cut off :code:`frequenciesabove`
-    :type low_pass: int
-    :param high_pass: The number to cut off :code:`frequenceisbelow`
-    :type high_pass: int
+#     :param data_emg: Poly5 file with the samples to work over
+#     :type data_emg: ~TMSiSDK.file_readers.Poly5Reader
+#     :param low_pass: The number to cut off :code:`frequenciesabove`
+#     :type low_pass: int
+#     :param high_pass: The number to cut off :code:`frequenceisbelow`
+#     :type high_pass: int
 
-    :returns: The bandpass filtered EMG sample data
-    :rtype: ~numpy.ndarray
-    """
-    sos = signal.butter(
-        3,
-        [low_pass, high_pass],
-        'bandpass',
-        fs=data_emg.sample_rate,
-        output='sos',
-    )
-    # sos (output parameter) is second order section  -> "stabilizes" ?
-    emg_filtered = signal.sosfiltfilt(sos, data_emg.samples)
-    return emg_filtered
-
-
-def emg_bandpass_butter_sample(
-    data_emg_samp,
-    low_pass,
-    high_pass,
-    sample_rate,
-    output='sos'
-):
-    """The paramemter taken in here is the Poly5 file.  Output is the
-    EMG after a bandpass as made here.
-
-    :param data_emg_samp: The array in the Poly5 or other sample
-    :type data_emg_samp: ~numpy.ndarray
-    :param low_pass: The number to cut off :code:`frequenciesabove`
-    :type low_pass: int
-    :param high_pass: The number to cut off :code:`frequenceisbelow`
-    :type high_pass: int
-    :param sample_rate: The number of samples per second i.e. Hertz
-    :type sample_rate: int
-    :param output: The type of sampling stabilizor
-    :type high_pass: str
-
-    :returns: The bandpass filtered EMG sample data
-    :rtype: ~numpy.ndarray
-    """
-    sos = signal.butter(
-        3,
-        [low_pass, high_pass],
-        'bandpass',
-        fs=sample_rate,
-        output='sos',
-    )
-    # sos (output parameter)is second order section  -> "stabilizes" ?
-    emg_filtered = signal.sosfiltfilt(sos, data_emg_samp)
-    return emg_filtered
+#     :returns: The bandpass filtered EMG sample data
+#     :rtype: ~numpy.ndarray
+#     """
+#     sos = signal.butter(
+#         3,
+#         [low_pass, high_pass],
+#         'bandpass',
+#         fs=data_emg.sample_rate,
+#         output='sos',
+#     )
+#     # sos (output parameter) is second order section  -> "stabilizes" ?
+#     emg_filtered = signal.sosfiltfilt(sos, data_emg.samples)
+#     return emg_filtered
 
 
-def bad_end_cutter(data_emg, percent_to_cut=7, tolerance_percent=10):
-    """This algorithm takes the end off of EMGs where the end is
-    radically altered, or if not radically altered cuts the last 10
-    values but returns only the array, not an altered Poly5.
+# def emg_bandpass_butter_sample(
+#     data_emg_samp,
+#     low_pass,
+#     high_pass,
+#     sample_rate,
+#     output='sos'
+# ):
+#     """The paramemter taken in here is the Poly5 file.  Output is the
+#     EMG after a bandpass as made here.
 
-    :param data_emg: A Poly5
-    :type data_emg: ~TMSiSDK.file_readers.Poly5Reader
-    :param percent_to_cut: Percentage to look at on the end
-    :type percent_to_cut: int
-    :param tolerance_percent: Percent variation tolerance to allow without
-        cutting
-    :type tolerance_percent: int
+#     :param data_emg_samp: The array in the Poly5 or other sample
+#     :type data_emg_samp: ~numpy.ndarray
+#     :param low_pass: The number to cut off :code:`frequenciesabove`
+#     :type low_pass: int
+#     :param high_pass: The number to cut off :code:`frequenceisbelow`
+#     :type high_pass: int
+#     :param sample_rate: The number of samples per second i.e. Hertz
+#     :type sample_rate: int
+#     :param output: The type of sampling stabilizor
+#     :type high_pass: str
 
-    :returns: The cut EMG sample data
-    :rtype: ~numpy.ndarray
-    """
-    sample = data_emg.samples
-    len_sample = len(data_emg.samples[0])
-
-    last_half = data_emg.samples[:, int(len_sample/2):]
-    percent = abs(int(percent_to_cut))
-    cut_number_last = int(((100 - percent)/100) * len_sample)
-    cut_number_last = int(((100 - percent)/100) * len_sample)
-    last_part = data_emg.samples[:, cut_number_last:]
-    leads = last_half.shape[0]
-    percent_off_list = []
-    for i in range(0, leads):
-        last_half_means = last_half[i].mean()
-        last_part_means = last_part[i].mean()
-        difference = abs(last_half_means - last_part_means)/last_half_means
-        percent_off_list.append(difference)
-    tolerance_list = []
-    for element in percent_off_list:
-        tolerance = tolerance_percent/100
-
-        if element >= tolerance:
-            booly = True
-            tolerance_list.append(booly)
-        else:
-            booly = False
-            tolerance_list.append(booly)
-
-    if True in tolerance_list:
-        sample_cut = sample[:, :cut_number_last]
-    else:
-        sample_cut = sample[:, :-10]
-
-    return sample_cut
+#     :returns: The bandpass filtered EMG sample data
+#     :rtype: ~numpy.ndarray
+#     """
+#     sos = signal.butter(
+#         3,
+#         [low_pass, high_pass],
+#         'bandpass',
+#         fs=sample_rate,
+#         output='sos',
+#     )
+#     # sos (output parameter)is second order section  -> "stabilizes" ?
+#     emg_filtered = signal.sosfiltfilt(sos, data_emg_samp)
+#     return emg_filtered
 
 
-def bad_end_cutter_for_samples(
-    data_emg,
-    percent_to_cut=7,
-    tolerance_percent=10
-):
-    """This algorithm takes the end off of EMGs where the end is
-    radically altered, or if not radically altered cuts the last 10
-    values but returns only the array.
+# def bad_end_cutter(data_emg, percent_to_cut=7, tolerance_percent=10):
+#     """This algorithm takes the end off of EMGs where the end is
+#     radically altered, or if not radically altered cuts the last 10
+#     values but returns only the array, not an altered Poly5.
 
-    :param data_emg: Array of samples
-    :type data_emg: ~numpy.ndarray
-    :param percent_to_cut: Percentage to look at on the end
-    :type percent_to_cut: int
-    :param tolerance_percent: Percent variation to allow without cutting
-    :type tolerance_percent: int
+#     :param data_emg: A Poly5
+#     :type data_emg: ~TMSiSDK.file_readers.Poly5Reader
+#     :param percent_to_cut: Percentage to look at on the end
+#     :type percent_to_cut: int
+#     :param tolerance_percent: Percent variation tolerance to allow without
+#         cutting
+#     :type tolerance_percent: int
 
-    :returns: The cut EMG sample data
-    :rtype: ~numpy.ndarray
-    """
-    sample = data_emg
-    len_sample = len(data_emg[0])
+#     :returns: The cut EMG sample data
+#     :rtype: ~numpy.ndarray
+#     """
+#     sample = data_emg.samples
+#     len_sample = len(data_emg.samples[0])
 
-    last_half = data_emg[:, int(len_sample/2):]
-    percent = abs(int(percent_to_cut))
-    cut_number_last = int(((100 - percent)/100) * len_sample)
+#     last_half = data_emg.samples[:, int(len_sample/2):]
+#     percent = abs(int(percent_to_cut))
+#     cut_number_last = int(((100 - percent)/100) * len_sample)
+#     cut_number_last = int(((100 - percent)/100) * len_sample)
+#     last_part = data_emg.samples[:, cut_number_last:]
+#     leads = last_half.shape[0]
+#     percent_off_list = []
+#     for i in range(0, leads):
+#         last_half_means = last_half[i].mean()
+#         last_part_means = last_part[i].mean()
+#         difference = abs(last_half_means - last_part_means)/last_half_means
+#         percent_off_list.append(difference)
+#     tolerance_list = []
+#     for element in percent_off_list:
+#         tolerance = tolerance_percent/100
 
-    last_part = data_emg[:, cut_number_last:]
-    leads = last_half.shape[0]
-    percent_off_list = []
-    # get rid of for loop, take advange of numpy array- next version
-    for i in range(leads):
-        last_half_means = last_half[i].mean()
-        last_part_means = last_part[i].mean()
-        difference = abs(last_half_means - last_part_means)/last_half_means
-        percent_off_list.append(difference)
-    tolerance = tolerance_percent / 100
-    if any(elt >= tolerance for elt in percent_off_list):
-        sample_cut = sample[:, :cut_number_last]
-    else:
-        sample_cut = sample[:, :-10]
+#         if element >= tolerance:
+#             booly = True
+#             tolerance_list.append(booly)
+#         else:
+#             booly = False
+#             tolerance_list.append(booly)
 
-    return sample_cut
+#     if True in tolerance_list:
+#         sample_cut = sample[:, :cut_number_last]
+#     else:
+#         sample_cut = sample[:, :-10]
 
-
-def bad_end_cutter_better(data_emg, percent_to_cut=7, tolerance_percent=10):
-    """This algorithm takes the end off of EMGs where the end is
-    radically altered, or if not radically altered cuts the last 10
-    values but returns only the array not an altered Poly5.
-
-    :param data_emg: A Poly5
-    :type data_emg: ~TMSiSDK.file_readers.Poly5Reader
-    :param percent_to_cut: Percentage to look at on the end
-    :type percent_to_cut: int
-    :param tolerance_percent: Percentage variation to allow without cut
-    :type tolerance_percent: int
-
-    :returns: The cut EMG sample data
-    :rtype: ~numpy.ndarray
-    """
-    sample = data_emg.samples
-    len_sample = len(data_emg.samples[0])
-
-    last_half = data_emg.samples[:, int(len_sample/2):]
-    percent = abs(int(percent_to_cut))
-    cut_number_last = int(((100 - percent)/100) * len_sample)
-
-    last_part = data_emg.samples[:, cut_number_last:]
-    leads = last_half.shape[0]
-    percent_off_list = []
-    # get rid of for loop, take advange of numpy array- next version
-    for i in range(leads):
-        last_half_means = last_half[i].mean()
-        last_part_means = last_part[i].mean()
-        difference = abs(last_half_means - last_part_means)/last_half_means
-        percent_off_list.append(difference)
-    tolerance = tolerance_percent / 100
-    if any(elt >= tolerance for elt in percent_off_list):
-        sample_cut = sample[:, :cut_number_last]
-    else:
-        sample_cut = sample[:, :-10]
-
-    return sample_cut
+#     return sample_cut
 
 
-def notch_filter(sample, sample_frequ, freq_to_pull, quality_factor_q):
-    """This is a filter designed to take out a specific frequency.  In
-    the EU in some data electrical cords can interfere at around 50
-    Hertz.  In some other locations the interference is at 60 Hertz.
-    The specificities of a local power grid may neccesitate notch
-    filtering.
+# def bad_end_cutter_for_samples(
+#     data_emg,
+#     percent_to_cut=7,
+#     tolerance_percent=10
+# ):
+#     """This algorithm takes the end off of EMGs where the end is
+#     radically altered, or if not radically altered cuts the last 10
+#     values but returns only the array.
 
-    :param sample: Percentage variation tolerance to allow without cutting
-    :type sample: int
-    :param sample_frequ: The frequency at which the sample was captured
-    :type sample_frequ: int
-    :param freq_to_pull: The frequency you desire to remove from the signal
-    :type freq_to_pull: int
-    :param quality_factor_q: How high the quality of the removal is
-    :type quality_factor_q: int
+#     :param data_emg: Array of samples
+#     :type data_emg: ~numpy.ndarray
+#     :param percent_to_cut: Percentage to look at on the end
+#     :type percent_to_cut: int
+#     :param tolerance_percent: Percent variation to allow without cutting
+#     :type tolerance_percent: int
 
-    :return: The filterered sample data
-    :rtype: ~numpy.ndarray
+#     :returns: The cut EMG sample data
+#     :rtype: ~numpy.ndarray
+#     """
+#     sample = data_emg
+#     len_sample = len(data_emg[0])
 
-    """
-    # create notch filter
-    # design a notch filter using signal.iirnotch
-    b_notch, a_notch = signal.iirnotch(
-        freq_to_pull,
-        quality_factor_q,
-        sample_frequ)
+#     last_half = data_emg[:, int(len_sample/2):]
+#     percent = abs(int(percent_to_cut))
+#     cut_number_last = int(((100 - percent)/100) * len_sample)
 
-    # make the output signal
-    output_signal = signal.filtfilt(b_notch, a_notch, sample)
-    return output_signal
+#     last_part = data_emg[:, cut_number_last:]
+#     leads = last_half.shape[0]
+#     percent_off_list = []
+#     # get rid of for loop, take advange of numpy array- next version
+#     for i in range(leads):
+#         last_half_means = last_half[i].mean()
+#         last_part_means = last_part[i].mean()
+#         difference = abs(last_half_means - last_part_means)/last_half_means
+#         percent_off_list.append(difference)
+#     tolerance = tolerance_percent / 100
+#     if any(elt >= tolerance for elt in percent_off_list):
+#         sample_cut = sample[:, :cut_number_last]
+#     else:
+#         sample_cut = sample[:, :-10]
+
+#     return sample_cut
+
+
+# def bad_end_cutter_better(data_emg, percent_to_cut=7, tolerance_percent=10):
+#     """This algorithm takes the end off of EMGs where the end is
+#     radically altered, or if not radically altered cuts the last 10
+#     values but returns only the array not an altered Poly5.
+
+#     :param data_emg: A Poly5
+#     :type data_emg: ~TMSiSDK.file_readers.Poly5Reader
+#     :param percent_to_cut: Percentage to look at on the end
+#     :type percent_to_cut: int
+#     :param tolerance_percent: Percentage variation to allow without cut
+#     :type tolerance_percent: int
+
+#     :returns: The cut EMG sample data
+#     :rtype: ~numpy.ndarray
+#     """
+#     sample = data_emg.samples
+#     len_sample = len(data_emg.samples[0])
+
+#     last_half = data_emg.samples[:, int(len_sample/2):]
+#     percent = abs(int(percent_to_cut))
+#     cut_number_last = int(((100 - percent)/100) * len_sample)
+
+#     last_part = data_emg.samples[:, cut_number_last:]
+#     leads = last_half.shape[0]
+#     percent_off_list = []
+#     # get rid of for loop, take advange of numpy array- next version
+#     for i in range(leads):
+#         last_half_means = last_half[i].mean()
+#         last_part_means = last_part[i].mean()
+#         difference = abs(last_half_means - last_part_means)/last_half_means
+#         percent_off_list.append(difference)
+#     tolerance = tolerance_percent / 100
+#     if any(elt >= tolerance for elt in percent_off_list):
+#         sample_cut = sample[:, :cut_number_last]
+#     else:
+#         sample_cut = sample[:, :-10]
+
+#     return sample_cut
+
+
+# def notch_filter(sample, sample_frequ, freq_to_pull, quality_factor_q):
+#     """This is a filter designed to take out a specific frequency.  In
+#     the EU in some data electrical cords can interfere at around 50
+#     Hertz.  In some other locations the interference is at 60 Hertz.
+#     The specificities of a local power grid may neccesitate notch
+#     filtering.
+
+#     :param sample: Percentage variation tolerance to allow without cutting
+#     :type sample: int
+#     :param sample_frequ: The frequency at which the sample was captured
+#     :type sample_frequ: int
+#     :param freq_to_pull: The frequency you desire to remove from the signal
+#     :type freq_to_pull: int
+#     :param quality_factor_q: How high the quality of the removal is
+#     :type quality_factor_q: int
+
+#     :return: The filterered sample data
+#     :rtype: ~numpy.ndarray
+
+#     """
+#     # create notch filter
+#     # design a notch filter using signal.iirnotch
+#     b_notch, a_notch = signal.iirnotch(
+#         freq_to_pull,
+#         quality_factor_q,
+#         sample_frequ)
+
+#     # make the output signal
+#     output_signal = signal.filtfilt(b_notch, a_notch, sample)
+#     return output_signal
 
 
 def show_my_power_spectrum(sample, sample_rate, upper_window):
@@ -335,24 +335,24 @@ def show_my_power_spectrum(sample, sample_rate, upper_window):
     return yf, xf
 
 
-def emg_highpass_butter(data_emg, cut_above, sample_rate):
-    """The parameter taken in here is the Poly5 file's samples or
-    another array.  Output is the EMG after a bandpass as made here.
+# def emg_highpass_butter(data_emg, cut_above, sample_rate):
+#     """The parameter taken in here is the Poly5 file's samples or
+#     another array.  Output is the EMG after a bandpass as made here.
 
-    :param data_emg: Samples from the EMG
-    :type data_emg: ~numpy.ndarray
-    :param cut_above: The number to cut off :code:`frequenceisbelow`
-    :type cut_above: int
-    :param sample_rate: The number of samples per second i.e. Hertz
-    :type sample_rate: int
+#     :param data_emg: Samples from the EMG
+#     :type data_emg: ~numpy.ndarray
+#     :param cut_above: The number to cut off :code:`frequenceisbelow`
+#     :type cut_above: int
+#     :param sample_rate: The number of samples per second i.e. Hertz
+#     :type sample_rate: int
 
-    :returns: The bandpass filtered EMG sample data
-    :rtype: ~numpy.ndarray
-    """
-    sos = signal.butter(3, cut_above, 'highpass', fs=sample_rate, output='sos')
-    # sos (output parameter)is second order section  -> "stabilizes" ?
-    emg_filtered = signal.sosfiltfilt(sos, data_emg)
-    return emg_filtered
+#     :returns: The bandpass filtered EMG sample data
+#     :rtype: ~numpy.ndarray
+#     """
+#     sos = signal.butter(3, cut_above, 'highpass', fs=sample_rate, output='sos')
+#     # sos (output parameter)is second order section  -> "stabilizes" ?
+#     emg_filtered = signal.sosfiltfilt(sos, data_emg)
+#     return emg_filtered
 
 
 def naive_rolling_rms(x, N):
@@ -433,121 +433,121 @@ def compute_ICA_two_comp(emg_samples):
     return component_0, component_1
 
 
-def pick_more_peaks_array(components_tuple):
-    """Here we have a function that takes a tuple with the two parts
-    of ICA, and finds the one with more peaks and anti-peaks.  The EMG
-    if without a final envelope will have more peaks
+# def pick_more_peaks_array(components_tuple):
+#     """Here we have a function that takes a tuple with the two parts
+#     of ICA, and finds the one with more peaks and anti-peaks.  The EMG
+#     if without a final envelope will have more peaks
 
-    .. note::
-        Data should not have been finally filtered to envelope level
+#     .. note::
+#         Data should not have been finally filtered to envelope level
 
-    :param components_tuple: tuple of two arrays representing different signals
-    :type components_tuple: Tuple[~numpy.ndarray, ~numpy.ndarray]
+#     :param components_tuple: tuple of two arrays representing different signals
+#     :type components_tuple: Tuple[~numpy.ndarray, ~numpy.ndarray]
 
-    :return: Array with more peaks (should usually be the EMG as
-        opposed to ECG)
-    :rtype: ~numpy.ndarray
-    """
-    c0 = components_tuple[0]
-    c1 = components_tuple[1]
-    low_border_c0 = (c0.max() - c0.mean())/4
-    peaks0, _0 = find_peaks(c0, height=low_border_c0, distance=10)
-    antipeaks0, anti_0 = find_peaks(
-        (c0*(-1)),
-        height=-low_border_c0,
-        distance=10)
-    low_border_c1 = (c1.max() - c1.mean())/4
-    peaks1, _1 = find_peaks(c1, height=low_border_c1, distance=10)
-    antipeaks1, anti_1 = find_peaks(
-        (c1*(-1)),
-        height=-low_border_c1,
-        distance=10,
-    )
+#     :return: Array with more peaks (should usually be the EMG as
+#         opposed to ECG)
+#     :rtype: ~numpy.ndarray
+#     """
+#     c0 = components_tuple[0]
+#     c1 = components_tuple[1]
+#     low_border_c0 = (c0.max() - c0.mean())/4
+#     peaks0, _0 = find_peaks(c0, height=low_border_c0, distance=10)
+#     antipeaks0, anti_0 = find_peaks(
+#         (c0*(-1)),
+#         height=-low_border_c0,
+#         distance=10)
+#     low_border_c1 = (c1.max() - c1.mean())/4
+#     peaks1, _1 = find_peaks(c1, height=low_border_c1, distance=10)
+#     antipeaks1, anti_1 = find_peaks(
+#         (c1*(-1)),
+#         height=-low_border_c1,
+#         distance=10,
+#     )
 
-    sum_peaks_0 = len(peaks0) + len(antipeaks0)
-    sum_peaks_1 = len(peaks1) + len(antipeaks1)
+#     sum_peaks_0 = len(peaks0) + len(antipeaks0)
+#     sum_peaks_1 = len(peaks1) + len(antipeaks1)
 
-    if sum_peaks_0 > sum_peaks_1:
-        emg_component = components_tuple[0]
-    elif sum_peaks_1 > sum_peaks_0:
-        emg_component = components_tuple[1]
-    else:
-        print("this is very strange data, please examine by hand")
-    return emg_component
-
-
-def pick_lowest_correlation_array(components_tuple, ecg_lead):
-    """Here we have a function that takes a tuple with the two parts
-    of ICA and the array containing the ECG recording, and finds the
-    ICA component with the lowest similarity to the ECG.
-    Data should not have been finally filtered to envelope level
-
-    :param components_tuple: tuple of two arrays representing different signals
-    :type components_tuple: Tuple[~numpy.ndarray, ~numpy.ndarray]
-
-    :param ecg_lead: array containing the ECG recording
-    :type ecg_lead: numpy.ndarray
-
-    :returns: Array with the lowest correlation coefficient
-     to the ECG lead (should usually be the EMG as opposed to ECG)
-    :rtype: ~numpy.ndarray
-    """
-    c0 = components_tuple[0]
-    c1 = components_tuple[1]
-
-    # create a tuple containing the data, each row is a variable,
-    # each column is an observation
-
-    corr_tuple = np.row_stack((ecg_lead, c0, c1))
-
-    # compute the correlation matrix
-    # the absolute value is used, because the ICA decomposition might
-    # produce a component with negative peaks. In this case
-    # the signals will be maximally negatively correlated
-
-    corr_matrix = abs(np.corrcoef(corr_tuple))
-
-    # get the component with the lowest correlation to ECG
-    # the matrix is symmetric, so we can check just the first row
-    # the first coefficient is the autocorrelation of the ECG lead,
-    # so we can check row 1 and 2
-
-    lowest_index = np.argmin(corr_matrix[0][1:])
-    emg_component = components_tuple[lowest_index]
-
-    return emg_component
+#     if sum_peaks_0 > sum_peaks_1:
+#         emg_component = components_tuple[0]
+#     elif sum_peaks_1 > sum_peaks_0:
+#         emg_component = components_tuple[1]
+#     else:
+#         print("this is very strange data, please examine by hand")
+#     return emg_component
 
 
-def pick_highest_correlation_array(components_tuple, ecg_lead):
-    """Here we have a function that takes a tuple with the two parts
-    of ICA and the array containing the ECG recording, and finds the
-    ICA component with the highest similarity to the ECG.
-    Data should not have been finally filtered to envelope level
+# def pick_lowest_correlation_array(components_tuple, ecg_lead):
+#     """Here we have a function that takes a tuple with the two parts
+#     of ICA and the array containing the ECG recording, and finds the
+#     ICA component with the lowest similarity to the ECG.
+#     Data should not have been finally filtered to envelope level
 
-    :param components_tuple: tuple of two arrays representing different signals
-    :type components_tuple: Tuple[~numpy.ndarray, ~numpy.ndarray]
+#     :param components_tuple: tuple of two arrays representing different signals
+#     :type components_tuple: Tuple[~numpy.ndarray, ~numpy.ndarray]
 
-    :param ecg_lead: array containing the ECG recording
-    :type ecg_lead: numpy.ndarray
+#     :param ecg_lead: array containing the ECG recording
+#     :type ecg_lead: numpy.ndarray
 
-    :returns: Array with the highest correlation coefficient
-     to the ECG lead (should usually be the  ECG)
-    :rtype: ~numpy.ndarray
-    """
-    c0 = components_tuple[0]
-    c1 = components_tuple[1]
-    corr_tuple = np.row_stack((ecg_lead, c0, c1))
-    corr_matrix = abs(np.corrcoef(corr_tuple))
+#     :returns: Array with the lowest correlation coefficient
+#      to the ECG lead (should usually be the EMG as opposed to ECG)
+#     :rtype: ~numpy.ndarray
+#     """
+#     c0 = components_tuple[0]
+#     c1 = components_tuple[1]
 
-    # get the component with the highest correlation to ECG
-    # the matrix is symmetric, so we can check just the first row
-    # the first coefficient is the autocorrelation of the ECG lead,
-    # so we can check row 1 and 2
+#     # create a tuple containing the data, each row is a variable,
+#     # each column is an observation
 
-    hi_index = np.argmax(corr_matrix[0][1:])
-    ecg_component = components_tuple[hi_index]
+#     corr_tuple = np.row_stack((ecg_lead, c0, c1))
 
-    return ecg_component
+#     # compute the correlation matrix
+#     # the absolute value is used, because the ICA decomposition might
+#     # produce a component with negative peaks. In this case
+#     # the signals will be maximally negatively correlated
+
+#     corr_matrix = abs(np.corrcoef(corr_tuple))
+
+#     # get the component with the lowest correlation to ECG
+#     # the matrix is symmetric, so we can check just the first row
+#     # the first coefficient is the autocorrelation of the ECG lead,
+#     # so we can check row 1 and 2
+
+#     lowest_index = np.argmin(corr_matrix[0][1:])
+#     emg_component = components_tuple[lowest_index]
+
+#     return emg_component
+
+
+# def pick_highest_correlation_array(components_tuple, ecg_lead):
+#     """Here we have a function that takes a tuple with the two parts
+#     of ICA and the array containing the ECG recording, and finds the
+#     ICA component with the highest similarity to the ECG.
+#     Data should not have been finally filtered to envelope level
+
+#     :param components_tuple: tuple of two arrays representing different signals
+#     :type components_tuple: Tuple[~numpy.ndarray, ~numpy.ndarray]
+
+#     :param ecg_lead: array containing the ECG recording
+#     :type ecg_lead: numpy.ndarray
+
+#     :returns: Array with the highest correlation coefficient
+#      to the ECG lead (should usually be the  ECG)
+#     :rtype: ~numpy.ndarray
+#     """
+#     c0 = components_tuple[0]
+#     c1 = components_tuple[1]
+#     corr_tuple = np.row_stack((ecg_lead, c0, c1))
+#     corr_matrix = abs(np.corrcoef(corr_tuple))
+
+#     # get the component with the highest correlation to ECG
+#     # the matrix is symmetric, so we can check just the first row
+#     # the first coefficient is the autocorrelation of the ECG lead,
+#     # so we can check row 1 and 2
+
+#     hi_index = np.argmax(corr_matrix[0][1:])
+#     ecg_component = components_tuple[hi_index]
+
+#     return ecg_component
 
 
 def working_pipeline_exp(our_chosen_file):
@@ -728,55 +728,55 @@ def entropy_scipy(sli, base=None):
     return entropy_count
 
 
-def compute_power_loss(
-    original_signal,
-    original_signal_sampling_frequency,
-    processed_signal,
-    processed_signal_sampling_frequency
-):
-    """This function computes the percentage of power loss after the
-    processing of a signal. Inputs include the original_signal (signal
-    before the processing), :code:`original_signal_sampling_frequency`
-    (sampling frequency of the signal before processing),
-    :code:`processed_signal` (signal after processing),
-    :code:`processed_signal_sampling_frequency` (sampling frequency of
-    the signal after processing).
+# def compute_power_loss(
+#     original_signal,
+#     original_signal_sampling_frequency,
+#     processed_signal,
+#     processed_signal_sampling_frequency
+# ):
+#     """This function computes the percentage of power loss after the
+#     processing of a signal. Inputs include the original_signal (signal
+#     before the processing), :code:`original_signal_sampling_frequency`
+#     (sampling frequency of the signal before processing),
+#     :code:`processed_signal` (signal after processing),
+#     :code:`processed_signal_sampling_frequency` (sampling frequency of
+#     the signal after processing).
 
-    Output is the percentage of power loss.
+#     Output is the percentage of power loss.
 
-    :param original_signal: Array.
-    :type  original_signal: ~numpy.ndarray
-    :param original_signal_sampling_frequency: Sampling freq. original signal
-    :type original_signal_sampling_frequency: int
-    :param processed_signal: Array.
-    :type  processed_signal: ~numpy.ndarray
-    :param processed_signal_sampling_frequency: Sampling frequency processed
-        signal
-    :type processed_signal_sampling_frequency: int
+#     :param original_signal: Array.
+#     :type  original_signal: ~numpy.ndarray
+#     :param original_signal_sampling_frequency: Sampling freq. original signal
+#     :type original_signal_sampling_frequency: int
+#     :param processed_signal: Array.
+#     :type  processed_signal: ~numpy.ndarray
+#     :param processed_signal_sampling_frequency: Sampling frequency processed
+#         signal
+#     :type processed_signal_sampling_frequency: int
 
-    :returns: Power loss
-    :rtype: float
-    """
-    nperseg = 1024
-    noverlap = 512
+#     :returns: Power loss
+#     :rtype: float
+#     """
+#     nperseg = 1024
+#     noverlap = 512
 
-    # power spectrum density of the original and
-    # processed signals using Welch method
-    Pxx_den_orig = signal.welch(  # as per Lu et al. 2009
-        original_signal,
-        original_signal_sampling_frequency,
-        nperseg=nperseg,
-        noverlap=noverlap,
-    )
-    Pxx_den_processed = signal.welch(
-        processed_signal,
-        processed_signal_sampling_frequency,
-        nperseg=nperseg,
-        noverlap=noverlap,)
-    # compute the percentage of power loss
-    power_loss = 100*(1-(np.sum(Pxx_den_processed)/np.sum(Pxx_den_orig)))
+#     # power spectrum density of the original and
+#     # processed signals using Welch method
+#     Pxx_den_orig = signal.welch(  # as per Lu et al. 2009
+#         original_signal,
+#         original_signal_sampling_frequency,
+#         nperseg=nperseg,
+#         noverlap=noverlap,
+#     )
+#     Pxx_den_processed = signal.welch(
+#         processed_signal,
+#         processed_signal_sampling_frequency,
+#         nperseg=nperseg,
+#         noverlap=noverlap,)
+#     # compute the percentage of power loss
+#     power_loss = 100*(1-(np.sum(Pxx_den_processed)/np.sum(Pxx_den_orig)))
 
-    return power_loss
+#     return power_loss
 
 
 def count_decision_array(decision_array):
@@ -998,115 +998,115 @@ def full_rolling_rms(x, N):
     return emg_rms
 
 
-def gating(
-    src_signal,
-    gate_peaks,
-    gate_width=205,
-    method=1,
-):
-    """
-    Eliminate peaks (e.g. QRS) from src_signal using gates
-    of width gate_width. The gate either filled by zeros or interpolation.
-    The filling method for the gate is encoded as follows:
-    0: Filled with zeros
-    1: Interpolation samples before and after
-    2: Fill with average of prior segment if exists
-    otherwise fill with post segment
-    3: Fill with running average of RMS (default)
+# def gating(
+#     src_signal,
+#     gate_peaks,
+#     gate_width=205,
+#     method=1,
+# ):
+#     """
+#     Eliminate peaks (e.g. QRS) from src_signal using gates
+#     of width gate_width. The gate either filled by zeros or interpolation.
+#     The filling method for the gate is encoded as follows:
+#     0: Filled with zeros
+#     1: Interpolation samples before and after
+#     2: Fill with average of prior segment if exists
+#     otherwise fill with post segment
+#     3: Fill with running average of RMS (default)
 
-    :param src_signal: Signal to process
-    :type src_signalsignal: ~numpy.ndarray
-    :param gate_peaks: list of individual peak index places to be gated
-    :type gate_peaks: ~list
-    :param gate_width: width of the gate
-    :type gate_width: int
-    :param method: filling method of gate
-    :type method: int
+#     :param src_signal: Signal to process
+#     :type src_signalsignal: ~numpy.ndarray
+#     :param gate_peaks: list of individual peak index places to be gated
+#     :type gate_peaks: ~list
+#     :param gate_width: width of the gate
+#     :type gate_width: int
+#     :param method: filling method of gate
+#     :type method: int
 
-    :returns: src_signal_gated, the gated result
-    :rtype: ~numpy.ndarray
-    """
-    src_signal_gated = copy.deepcopy(src_signal)
-    max_sample = src_signal_gated.shape[0]
-    half_gate_width = gate_width // 2
-    if method == 0:
-        # Method 0: Fill with zeros
-        # TODO: can rewrite with slices from numpy irange to be more efficient
-        gate_samples = []
-        for i, peak in enumerate(gate_peaks):
-            for k in range(
-                max(0, peak - half_gate_width),
-                min(max_sample, peak + half_gate_width),
-            ):
-                gate_samples.append(k)
+#     :returns: src_signal_gated, the gated result
+#     :rtype: ~numpy.ndarray
+#     """
+#     src_signal_gated = copy.deepcopy(src_signal)
+#     max_sample = src_signal_gated.shape[0]
+#     half_gate_width = gate_width // 2
+#     if method == 0:
+#         # Method 0: Fill with zeros
+#         # TODO: can rewrite with slices from numpy irange to be more efficient
+#         gate_samples = []
+#         for i, peak in enumerate(gate_peaks):
+#             for k in range(
+#                 max(0, peak - half_gate_width),
+#                 min(max_sample, peak + half_gate_width),
+#             ):
+#                 gate_samples.append(k)
 
-        src_signal_gated[gate_samples] = 0
-    elif method == 1:
-        # Method 1: Fill with interpolation pre- and post gate sample
-        # TODO: rewrite with numpy interpolation for efficiency
-        for i, peak in enumerate(gate_peaks):
-            pre_ave_emg = src_signal[peak-half_gate_width-1]
+#         src_signal_gated[gate_samples] = 0
+#     elif method == 1:
+#         # Method 1: Fill with interpolation pre- and post gate sample
+#         # TODO: rewrite with numpy interpolation for efficiency
+#         for i, peak in enumerate(gate_peaks):
+#             pre_ave_emg = src_signal[peak-half_gate_width-1]
 
-            if (peak + half_gate_width + 1) < src_signal_gated.shape[0]:
-                post_ave_emg = src_signal[peak+half_gate_width+1]
-            else:
-                post_ave_emg = 0
+#             if (peak + half_gate_width + 1) < src_signal_gated.shape[0]:
+#                 post_ave_emg = src_signal[peak+half_gate_width+1]
+#             else:
+#                 post_ave_emg = 0
 
-            k_start = max(0, peak-half_gate_width)
-            k_end = min(
-                peak+half_gate_width, src_signal_gated.shape[0]
-            )
-            for k in range(k_start, k_end):
-                frac = (k - peak + half_gate_width)/gate_width
-                loup = (1 - frac) * pre_ave_emg + frac * post_ave_emg
-                src_signal_gated[k] = loup
+#             k_start = max(0, peak-half_gate_width)
+#             k_end = min(
+#                 peak+half_gate_width, src_signal_gated.shape[0]
+#             )
+#             for k in range(k_start, k_end):
+#                 frac = (k - peak + half_gate_width)/gate_width
+#                 loup = (1 - frac) * pre_ave_emg + frac * post_ave_emg
+#                 src_signal_gated[k] = loup
 
-    elif method == 2:
-        # Method 2: Fill with window length mean over prior section
-        # ..._____|_______|_______|XXXXXXX|XXXXXXX|_____...
-        #         ^               ^- gate start   ^- gate end
-        #         - peak - half_gate_width * 3 (replacer)
+#     elif method == 2:
+#         # Method 2: Fill with window length mean over prior section
+#         # ..._____|_______|_______|XXXXXXX|XXXXXXX|_____...
+#         #         ^               ^- gate start   ^- gate end
+#         #         - peak - half_gate_width * 3 (replacer)
 
-        for i, peak in enumerate(gate_peaks):
-            start = peak - half_gate_width * 3
-            if start < 0:
-                start = peak + half_gate_width
-            end = start + gate_width
-            pre_ave_emg = np.nanmean(src_signal[start:end])
+#         for i, peak in enumerate(gate_peaks):
+#             start = peak - half_gate_width * 3
+#             if start < 0:
+#                 start = peak + half_gate_width
+#             end = start + gate_width
+#             pre_ave_emg = np.nanmean(src_signal[start:end])
 
-            k_start = max(0, peak - half_gate_width)
-            k_end = min(peak + half_gate_width, src_signal_gated.shape[0])
-            for k in range(k_start, k_end):
-                src_signal_gated[k] = pre_ave_emg
+#             k_start = max(0, peak - half_gate_width)
+#             k_end = min(peak + half_gate_width, src_signal_gated.shape[0])
+#             for k in range(k_start, k_end):
+#                 src_signal_gated[k] = pre_ave_emg
 
-    elif method == 3:
-        # Method 3: Fill with moving average over RMS
-        gate_samples = []
-        for i, peak in enumerate(gate_peaks):
-            for k in range(
-                max([0, int(peak-gate_width/2)]),
-                min([max_sample, int(peak+gate_width/2)])
-            ):
-                gate_samples.append(k)
+#     elif method == 3:
+#         # Method 3: Fill with moving average over RMS
+#         gate_samples = []
+#         for i, peak in enumerate(gate_peaks):
+#             for k in range(
+#                 max([0, int(peak-gate_width/2)]),
+#                 min([max_sample, int(peak+gate_width/2)])
+#             ):
+#                 gate_samples.append(k)
 
-        src_signal_gated_base = copy.deepcopy(src_signal_gated)
-        src_signal_gated_base[gate_samples] = np.NaN
-        src_signal_gated_rms = full_rolling_rms(
-            src_signal_gated_base,
-            gate_width,)
+#         src_signal_gated_base = copy.deepcopy(src_signal_gated)
+#         src_signal_gated_base[gate_samples] = np.NaN
+#         src_signal_gated_rms = full_rolling_rms(
+#             src_signal_gated_base,
+#             gate_width,)
 
-        for i, peak in enumerate(gate_peaks):
-            k_start = max([0, int(peak-gate_width/2)])
-            k_end = min([int(peak+gate_width/2), max_sample])
+#         for i, peak in enumerate(gate_peaks):
+#             k_start = max([0, int(peak-gate_width/2)])
+#             k_end = min([int(peak+gate_width/2), max_sample])
 
-            for k in range(k_start, k_end):
-                leftf = max([0, int(k-1.5*gate_width)])
-                rightf = min([int(k+1.5*gate_width), max_sample])
-                src_signal_gated[k] = np.nanmean(
-                    src_signal_gated_rms[leftf:rightf]
-                )
+#             for k in range(k_start, k_end):
+#                 leftf = max([0, int(k-1.5*gate_width)])
+#                 rightf = min([int(k+1.5*gate_width), max_sample])
+#                 src_signal_gated[k] = np.nanmean(
+#                     src_signal_gated_rms[leftf:rightf]
+#                 )
 
-    return src_signal_gated
+#     return src_signal_gated
 
 
 def merge(left, right):
@@ -1168,27 +1168,27 @@ def hi_envelope(our_signal, dmax=24):
     return smoothed_interped
 
 
-def compute_ICA_two_comp_multi(emg_samples):
-    """A function that performs an independant component analysis
-    (ICA) meant for EMG data that includes stacked arrays,
-    there should be at least two arrays but there can be more.
+# def compute_ICA_two_comp_multi(emg_samples):
+#     """A function that performs an independant component analysis
+#     (ICA) meant for EMG data that includes stacked arrays,
+#     there should be at least two arrays but there can be more.
 
-    :param emg_samples: Original signal array with three or more layers
-    :type emg_samples: ~numpy.ndarray
+#     :param emg_samples: Original signal array with three or more layers
+#     :type emg_samples: ~numpy.ndarray
 
-    :returns: Two arrays of independent components (ECG-like and EMG)
-    :rtype: ~numpy.ndarray
-    """
-    all_component_numbers = list(range(emg_samples.shape[0]))
-    list_to_c = []
-    for i in all_component_numbers:
-        list_to_c.append(emg_samples[i])
-    X = np.column_stack(list_to_c)
-    ica = FastICA(n_components=2, random_state=1)
-    S = ica.fit_transform(X)
-    component_0 = S.T[0]
-    component_1 = S.T[1]
-    return component_0, component_1
+#     :returns: Two arrays of independent components (ECG-like and EMG)
+#     :rtype: ~numpy.ndarray
+#     """
+#     all_component_numbers = list(range(emg_samples.shape[0]))
+#     list_to_c = []
+#     for i in all_component_numbers:
+#         list_to_c.append(emg_samples[i])
+#     X = np.column_stack(list_to_c)
+#     ica = FastICA(n_components=2, random_state=1)
+#     S = ica.fit_transform(X)
+#     component_0 = S.T[0]
+#     component_1 = S.T[1]
+#     return component_0, component_1
 
 
 def scale_arrays(array, maximumn, minimumn):
@@ -1477,59 +1477,59 @@ def distance_matrix(array_a, array_b):
     return distances
 
 
-def helper_lowpass(cutoff, fs, order=5):
-    """
-    This is a helper function inside the butter_lowpass_filter function.
-    """
-    return butter(order, cutoff, fs=fs, btype='low', analog=False)
+# def helper_lowpass(cutoff, fs, order=5):
+#     """
+#     This is a helper function inside the butter_lowpass_filter function.
+#     """
+#     return butter(order, cutoff, fs=fs, btype='low', analog=False)
 
 
-def emg_lowpass_butter(array, cutoff, fs, order=5):
-    """
-    This is a lowpass filter of butterworth design.
+# def emg_lowpass_butter(array, cutoff, fs, order=5):
+#     """
+#     This is a lowpass filter of butterworth design.
 
-    :param array: 1d signal array usually of EMG
-    :type array: ~numpy.ndarray
-    :param cutoff: frequency above which to filter out
-    :type cutoff: int
-    :param fs: frequency array sampled at in Hertz
-    :type fs: int
-    :param order: order of the filter
-    :type order: int
+#     :param array: 1d signal array usually of EMG
+#     :type array: ~numpy.ndarray
+#     :param cutoff: frequency above which to filter out
+#     :type cutoff: int
+#     :param fs: frequency array sampled at in Hertz
+#     :type fs: int
+#     :param order: order of the filter
+#     :type order: int
 
-    :returns: signal_filtered
-    :rtype: ~numpy.ndarray
-    """
-    b, a = helper_lowpass(cutoff, fs, order=order)
-    signal_filtered = lfilter(b, a, array)
-    return signal_filtered
+#     :returns: signal_filtered
+#     :rtype: ~numpy.ndarray
+#     """
+#     b, a = helper_lowpass(cutoff, fs, order=order)
+#     signal_filtered = lfilter(b, a, array)
+#     return signal_filtered
 
 
-def find_peaks_in_ecg_signal(ecg_signal, lower_border_percent=50):
-    """
-    This function assumes you have isolated an ecg-like signal with
-    QRS peaks "higher" (or lower) than ST waves.
-    In this case it can be applied to return an array of
-    ECG peak locations. NB: This function assumes that the ECG
-    signal has already been through a bandpass or low-pass filter
-    or has little baseline drift.
+# def find_peaks_in_ecg_signal(ecg_signal, lower_border_percent=50):
+#     """
+#     This function assumes you have isolated an ecg-like signal with
+#     QRS peaks "higher" (or lower) than ST waves.
+#     In this case it can be applied to return an array of
+#     ECG peak locations. NB: This function assumes that the ECG
+#     signal has already been through a bandpass or low-pass filter
+#     or has little baseline drift.
 
-    :param ecg_signal: frequency array sampled at in Hertz
-    :type ecg_signal: ~numpy.ndarray
-    :param low_border_percent: percentage max below which no peaks expected
-    :type low_border_percent: int
+#     :param ecg_signal: frequency array sampled at in Hertz
+#     :type ecg_signal: ~numpy.ndarray
+#     :param low_border_percent: percentage max below which no peaks expected
+#     :type low_border_percent: int
 
-    :returns: tuple first element peak locations, next a dictionary of info
-    :rtype: tuple
+#     :returns: tuple first element peak locations, next a dictionary of info
+#     :rtype: tuple
 
-    """
-    ecg_signal = abs(ecg_signal)
-    max_peak = ecg_signal.max() - ecg_signal.min()
-    set_ecg_peaks = find_peaks(
-        ecg_signal,
-        prominence=(max_peak*lower_border_percent/100, max_peak)
-    )
-    return set_ecg_peaks
+#     """
+#     ecg_signal = abs(ecg_signal)
+#     max_peak = ecg_signal.max() - ecg_signal.min()
+#     set_ecg_peaks = find_peaks(
+#         ecg_signal,
+#         prominence=(max_peak*lower_border_percent/100, max_peak)
+#     )
+#     return set_ecg_peaks
 
 
 def variability_maker(
