+++ tests/test.py
@@ -18,31 +18,31 @@ from resurfemg.config.config import hash_it_up_right_all
 from resurfemg.multi_lead_type import compute_ICA_n_comp
 from resurfemg.multi_lead_type import compute_ICA_n_comp_selective_zeroing
 from resurfemg.multi_lead_type import compute_ICA_two_comp_selective
-from resurfemg.multi_lead_type import pick_highest_correlation_array_multi
+# from resurfemg.multi_lead_type import pick_highest_correlation_array_multi
 from resurfemg.multi_lead_type import working_pipe_multi
 from resurfemg.multi_lead_type import working_pipeline_pre_ml_multi
 # helper_functions
-from resurfemg.helper_functions import bad_end_cutter
-from resurfemg.helper_functions import bad_end_cutter_better
-from resurfemg.helper_functions import bad_end_cutter_for_samples
+# from resurfemg.helper_functions import bad_end_cutter
+# from resurfemg.helper_functions import bad_end_cutter_better
+# from resurfemg.helper_functions import bad_end_cutter_for_samples
 from resurfemg.helper_functions import count_decision_array
-from resurfemg.helper_functions import emg_bandpass_butter
-from resurfemg.helper_functions import emg_bandpass_butter_sample
-from resurfemg.helper_functions import notch_filter
-from resurfemg.helper_functions import naive_rolling_rms
-from resurfemg.helper_functions import vect_naive_rolling_rms
+# from resurfemg.helper_functions import emg_bandpass_butter
+# from resurfemg.helper_functions import emg_bandpass_butter_sample
+# from resurfemg.helper_functions import notch_filter
+# from resurfemg.helper_functions import naive_rolling_rms
+# from resurfemg.helper_functions import vect_naive_rolling_rms
 from resurfemg.helper_functions import pick_more_peaks_array
-from resurfemg.helper_functions import pick_lowest_correlation_array
+# from resurfemg.helper_functions import pick_lowest_correlation_array
 from resurfemg.helper_functions import zero_one_for_jumps_base
-from resurfemg.helper_functions import compute_ICA_two_comp
-from resurfemg.helper_functions import compute_ICA_two_comp_multi
+# from resurfemg.helper_functions import compute_ICA_two_comp
+# from resurfemg.helper_functions import compute_ICA_two_comp_multi
 from resurfemg.helper_functions import working_pipeline_exp
 from resurfemg.helper_functions import entropical
 from resurfemg.helper_functions import entropy_scipy
-from resurfemg.helper_functions import smooth_for_baseline
-from resurfemg.helper_functions import smooth_for_baseline_with_overlay
+# from resurfemg.helper_functions import smooth_for_baseline
+# from resurfemg.helper_functions import smooth_for_baseline_with_overlay
 from resurfemg.helper_functions import relative_levenshtein
-from resurfemg.helper_functions import gating
+# from resurfemg.helper_functions import gating
 from resurfemg.helper_functions import scale_arrays
 from resurfemg.helper_functions import pseudo_slope
 from resurfemg.helper_functions import area_under_curve
@@ -50,8 +50,8 @@ from resurfemg.helper_functions import simple_area_under_curve
 from resurfemg.helper_functions import times_under_curve
 from resurfemg.helper_functions import find_peak_in_breath
 from resurfemg.helper_functions import distance_matrix
-from resurfemg.helper_functions import emg_lowpass_butter
-from resurfemg.helper_functions import find_peaks_in_ecg_signal
+# from resurfemg.helper_functions import emg_lowpass_butter
+# from resurfemg.helper_functions import find_peaks_in_ecg_signal
 from resurfemg.helper_functions import variability_maker
 # config
 from resurfemg.config.config import Config
@@ -97,52 +97,52 @@ class TestHashMethods(unittest.TestCase):
 
 class TestComponentPickingMethods(unittest.TestCase):
 
-    def test_pick_more_peaks_array(self):
-        sample_read= Poly5Reader(sample_emg)
-        sample_emg_filtered = emg_bandpass_butter(sample_read, 1, 10)
-        sample_emg_filtered[1]= sample_emg_filtered[0]*1.5
-        sample_emg_filtered[2]= sample_emg_filtered[0]*1.7
-        components = compute_ICA_two_comp(sample_emg_filtered)
-        emg = pick_more_peaks_array(components)
-        self.assertEqual(
-            (len(emg)),
-            len(components[0]) ,
-        )
-    def test_pick_lowest_correlation_array(self):
-        sample_read= Poly5Reader(sample_emg)
-        sample_emg_filtered = emg_bandpass_butter(sample_read, 1, 10)
-        sample_emg_filtered[1] = sample_emg_filtered[0]
-        sample_emg_filtered[2] = sample_emg_filtered[0]*0.7
-        sample_emg_filtered[2,20] = 0
-        sample_emg_filtered[2,40] = 0
-        sample_emg_filtered[2,80] = 0
-        sample_emg_filtered[2,21] = 100
-        sample_emg_filtered[2,42] = 100
-        sample_emg_filtered[2,81] = 100
-        components = sample_emg_filtered[1], sample_emg_filtered[2]
-        emg = pick_lowest_correlation_array(components, sample_emg_filtered[0])
-        self.assertEqual(
-            sum(emg),
-            sum(sample_emg_filtered[2]) ,
-        )
+    # def test_pick_more_peaks_array(self):
+    #     sample_read= Poly5Reader(sample_emg)
+    #     sample_emg_filtered = emg_bandpass_butter(sample_read, 1, 10)
+    #     sample_emg_filtered[1]= sample_emg_filtered[0]*1.5
+    #     sample_emg_filtered[2]= sample_emg_filtered[0]*1.7
+    #     components = compute_ICA_two_comp(sample_emg_filtered)
+    #     emg = pick_more_peaks_array(components)
+    #     self.assertEqual(
+    #         (len(emg)),
+    #         len(components[0]) ,
+    #     )
+    # def test_pick_lowest_correlation_array(self):
+    #     sample_read= Poly5Reader(sample_emg)
+    #     sample_emg_filtered = emg_bandpass_butter(sample_read, 1, 10)
+    #     sample_emg_filtered[1] = sample_emg_filtered[0]
+    #     sample_emg_filtered[2] = sample_emg_filtered[0]*0.7
+    #     sample_emg_filtered[2,20] = 0
+    #     sample_emg_filtered[2,40] = 0
+    #     sample_emg_filtered[2,80] = 0
+    #     sample_emg_filtered[2,21] = 100
+    #     sample_emg_filtered[2,42] = 100
+    #     sample_emg_filtered[2,81] = 100
+    #     components = sample_emg_filtered[1], sample_emg_filtered[2]
+    #     emg = pick_lowest_correlation_array(components, sample_emg_filtered[0])
+    #     self.assertEqual(
+    #         sum(emg),
+    #         sum(sample_emg_filtered[2]) ,
+    #     )
 
-    def test_pick_highest_correlation_array_multi(self):
-        sample_read= Poly5Reader(sample_emg)
-        sample_emg_filtered = emg_bandpass_butter(sample_read, 1, 10)
-        sample_emg_filtered[1] = sample_emg_filtered[0]
-        sample_emg_filtered[2] = sample_emg_filtered[0]*0.7
-        sample_emg_filtered[2,20] = 0
-        sample_emg_filtered[2,40] = 0
-        sample_emg_filtered[2,80] = 0
-        sample_emg_filtered[2,21] = 100
-        sample_emg_filtered[2,42] = 100
-        sample_emg_filtered[2,81] = 100
-        components = np.row_stack((sample_emg_filtered[1], sample_emg_filtered[2]))
-        emg = pick_highest_correlation_array_multi(components, sample_emg_filtered[0])
-        self.assertEqual(
-            sum(components[emg]),
-            sum(sample_emg_filtered[1]),
-        )
+    # def test_pick_highest_correlation_array_multi(self):
+    #     sample_read= Poly5Reader(sample_emg)
+    #     sample_emg_filtered = emg_bandpass_butter(sample_read, 1, 10)
+    #     sample_emg_filtered[1] = sample_emg_filtered[0]
+    #     sample_emg_filtered[2] = sample_emg_filtered[0]*0.7
+    #     sample_emg_filtered[2,20] = 0
+    #     sample_emg_filtered[2,40] = 0
+    #     sample_emg_filtered[2,80] = 0
+    #     sample_emg_filtered[2,21] = 100
+    #     sample_emg_filtered[2,42] = 100
+    #     sample_emg_filtered[2,81] = 100
+    #     components = np.row_stack((sample_emg_filtered[1], sample_emg_filtered[2]))
+    #     emg = pick_highest_correlation_array_multi(components, sample_emg_filtered[0])
+    #     self.assertEqual(
+    #         sum(components[emg]),
+    #         sum(sample_emg_filtered[1]),
+    #     )
 
 
 class TestEntropyMethods(unittest.TestCase):
@@ -182,67 +182,67 @@ class TestVariabilityMethods(unittest.TestCase):
     
 class TestFilteringMethods(unittest.TestCase):
 
-    def test_emg_band_pass_butter(self):
-        sample_read= Poly5Reader(sample_emg)
-        sample_emg_filtered = emg_bandpass_butter(sample_read, 1, 10)
-        self.assertEqual(
-            (len(sample_emg_filtered[0])),
-            len(sample_read.samples[0]) ,
-        )
-    def test_emg_band_pass_butter_sample(self):
-        sample_read= Poly5Reader(sample_emg)
-        sample_emg_filtered = emg_bandpass_butter_sample(sample_read.samples, 1, 10, 2048)
-        self.assertEqual(
-            (len(sample_emg_filtered[0])),
-            len(sample_read.samples[0]) ,
-        )
+    # def test_emg_band_pass_butter(self):
+    #     sample_read= Poly5Reader(sample_emg)
+    #     sample_emg_filtered = emg_bandpass_butter(sample_read, 1, 10)
+    #     self.assertEqual(
+    #         (len(sample_emg_filtered[0])),
+    #         len(sample_read.samples[0]) ,
+    #     )
+    # def test_emg_band_pass_butter_sample(self):
+    #     sample_read= Poly5Reader(sample_emg)
+    #     sample_emg_filtered = emg_bandpass_butter_sample(sample_read.samples, 1, 10, 2048)
+    #     self.assertEqual(
+    #         (len(sample_emg_filtered[0])),
+    #         len(sample_read.samples[0]) ,
+    #     )
     
-    def test_emg_lowpass_butter(self):
-        sample_read= Poly5Reader(sample_emg)
-        sample_emg_filtered = emg_lowpass_butter(sample_read.samples, 5, 2048)
-        self.assertEqual(
-            (len(sample_emg_filtered[0])),
-            len(sample_read.samples[0]) ,
-        )
+    # def test_emg_lowpass_butter(self):
+    #     sample_read= Poly5Reader(sample_emg)
+    #     sample_emg_filtered = emg_lowpass_butter(sample_read.samples, 5, 2048)
+    #     self.assertEqual(
+    #         (len(sample_emg_filtered[0])),
+    #         len(sample_read.samples[0]) ,
+    #     )
 
-    def test_notch_filter(self):
-        sample_read= Poly5Reader(sample_emg)
-        sample_emg_filtered = notch_filter(sample_read.samples, 2048, 80,2)
-        self.assertEqual(
-            (len(sample_emg_filtered[0])),
-            len(sample_read.samples[0]) ,
-        )
+    # def test_notch_filter(self):
+    #     sample_read= Poly5Reader(sample_emg)
+    #     sample_emg_filtered = notch_filter(sample_read.samples, 2048, 80,2)
+    #     self.assertEqual(
+    #         (len(sample_emg_filtered[0])),
+    #         len(sample_read.samples[0]) ,
+    #     )
 
-    def test_naive_rolling_rms(self):
-        sample_read= Poly5Reader(sample_emg)
-        sample_emg_filtered = naive_rolling_rms(sample_read.samples[0], 10)
-        self.assertNotEqual(
-            (len(sample_emg_filtered)),
-            len(sample_read.samples[0]) ,
-        )
+    # def test_naive_rolling_rms(self):
+    #     sample_read= Poly5Reader(sample_emg)
+    #     sample_emg_filtered = naive_rolling_rms(sample_read.samples[0], 10)
+    #     self.assertNotEqual(
+    #         (len(sample_emg_filtered)),
+    #         len(sample_read.samples[0]) ,
+    #     )
     
-    def test_vect_naive_rolling_rms(self):
-        sample_read= Poly5Reader(sample_emg)
-        sample_emg_filtered = vect_naive_rolling_rms(sample_read.samples[0], 10)
-        self.assertNotEqual(
-            (len(sample_emg_filtered)),
-            len(sample_read.samples[0]) ,
-        )
+    # def test_vect_naive_rolling_rms(self):
+    #     sample_read= Poly5Reader(sample_emg)
+    #     sample_emg_filtered = vect_naive_rolling_rms(sample_read.samples[0], 10)
+    #     self.assertNotEqual(
+    #         (len(sample_emg_filtered)),
+    #         len(sample_read.samples[0]) ,
+    #     )
 
 
 
 class TestPickingMethods(unittest.TestCase):
 
-    def test_compute_ICA_two_comp(self):
-        sample_read= Poly5Reader(sample_emg)
-        sample_emg_filtered = emg_bandpass_butter(sample_read, 1, 10)
-        sample_emg_filtered[1]= sample_emg_filtered[0]*1.5
-        sample_emg_filtered[2]= sample_emg_filtered[0]*1.7
-        components = compute_ICA_two_comp(sample_emg_filtered)
-        self.assertEqual(
-            (len(components[1])),
-            len(components[0]) ,
-        )
+    # def test_compute_ICA_two_comp(self):
+    #     sample_read= Poly5Reader(sample_emg)
+    #     sample_emg_filtered = emg_bandpass_butter(sample_read, 1, 10)
+    #     sample_emg_filtered[1]= sample_emg_filtered[0]*1.5
+    #     sample_emg_filtered[2]= sample_emg_filtered[0]*1.7
+    #     components = compute_ICA_two_comp(sample_emg_filtered)
+    #     self.assertEqual(
+    #         (len(components[1])),
+    #         len(components[0]) ,
+    #     )
 
     def test_compute_ICA_n_comp(self):
         sample_read= Poly5Reader(sample_emg)
@@ -268,16 +268,16 @@ class TestPickingMethods(unittest.TestCase):
             6,
         )
 
-    def test_compute_ICA_two_comp_multi(self):
-        sample_read= Poly5Reader(sample_emg)
-        sample_emg_filtered = emg_bandpass_butter(sample_read, 1, 10)
-        sample_emg_filtered[1]= sample_emg_filtered[0]*1.5
-        sample_emg_filtered[2]= sample_emg_filtered[0]*1.7
-        components = compute_ICA_two_comp_multi(sample_emg_filtered)
-        self.assertEqual(
-            (len(components)),
-            2 ,
-        )
+    # def test_compute_ICA_two_comp_multi(self):
+    #     sample_read= Poly5Reader(sample_emg)
+    #     sample_emg_filtered = emg_bandpass_butter(sample_read, 1, 10)
+    #     sample_emg_filtered[1]= sample_emg_filtered[0]*1.5
+    #     sample_emg_filtered[2]= sample_emg_filtered[0]*1.7
+    #     components = compute_ICA_two_comp_multi(sample_emg_filtered)
+    #     self.assertEqual(
+    #         (len(components)),
+    #         2 ,
+    #     )
 
     def test_compute_ICA_two_comp_selective(self):
         sample_read= Poly5Reader(sample_emg)
@@ -322,29 +322,29 @@ class TestPipelineMethods(unittest.TestCase):
         )
 
 
-class TestCuttingingMethods(unittest.TestCase):
+# class TestCuttingingMethods(unittest.TestCase):
 
-    def test_emg_bad_end_cutter(self):
-        sample_ready= Poly5Reader(sample_emg)
-        sample_emg_cut = bad_end_cutter(sample_ready, 1, 10)
-        self.assertNotEqual(
-            (len(sample_emg_cut[0])),
-            len(sample_ready.samples[0]),
-        )
-    def test_emg_bad_end_cutter_for_samples(self):
-        sample_read= Poly5Reader(sample_emg)
-        sample_emg_cut = bad_end_cutter_for_samples(sample_read.samples, 1, 10)
-        self.assertNotEqual(
-            (len(sample_emg_cut[0])),
-            len(sample_read.samples[0]) ,
-        )
-    def test_emg_bad_end_cutter_better(self):
-        sample_read= Poly5Reader(sample_emg)
-        sample_emg_cut = bad_end_cutter_better(sample_read, 1, 10)
-        self.assertNotEqual(
-            (len(sample_emg_cut[0])),
-            len(sample_read.samples[0]) ,
-        )
+#     def test_emg_bad_end_cutter(self):
+#         sample_ready= Poly5Reader(sample_emg)
+#         sample_emg_cut = bad_end_cutter(sample_ready, 1, 10)
+#         self.assertNotEqual(
+#             (len(sample_emg_cut[0])),
+#             len(sample_ready.samples[0]),
+#         )
+#     def test_emg_bad_end_cutter_for_samples(self):
+#         sample_read= Poly5Reader(sample_emg)
+#         sample_emg_cut = bad_end_cutter_for_samples(sample_read.samples, 1, 10)
+#         self.assertNotEqual(
+#             (len(sample_emg_cut[0])),
+#             len(sample_read.samples[0]) ,
+#         )
+#     def test_emg_bad_end_cutter_better(self):
+#         sample_read= Poly5Reader(sample_emg)
+#         sample_emg_cut = bad_end_cutter_better(sample_read, 1, 10)
+#         self.assertNotEqual(
+#             (len(sample_emg_cut[0])),
+#             len(sample_read.samples[0]) ,
+#         )
 
 class TestVentCompareMethods(unittest.TestCase):
 
@@ -356,56 +356,56 @@ class TestVentCompareMethods(unittest.TestCase):
         our_result13 = (relative_levenshtein(array1,array3))
         self.assertEqual(our_result12, our_result13)
 
-class TestGating(unittest.TestCase):
-    sample_read= Poly5Reader(sample_emg)
-    sample_emg_filtered = -emg_bandpass_butter(sample_read, 1, 500)
-    sample_emg_filtered = sample_emg_filtered[:30*2048]
-    ecg_peaks, _  = scipy.signal.find_peaks(sample_emg_filtered[0, :])
+# class TestGating(unittest.TestCase):
+#     sample_read= Poly5Reader(sample_emg)
+#     sample_emg_filtered = -emg_bandpass_butter(sample_read, 1, 500)
+#     sample_emg_filtered = sample_emg_filtered[:30*2048]
+#     ecg_peaks, _  = scipy.signal.find_peaks(sample_emg_filtered[0, :])
 
-    def test_gating_method_0(self):
-        ecg_gated_0 = gating(self.sample_emg_filtered[0, :], self.ecg_peaks, gate_width=205, method=0)
+#     def test_gating_method_0(self):
+#         ecg_gated_0 = gating(self.sample_emg_filtered[0, :], self.ecg_peaks, gate_width=205, method=0)
 
-        self.assertEqual(
-            (len(self.sample_emg_filtered[0])),
-            len(ecg_gated_0) ,
-        )
+#         self.assertEqual(
+#             (len(self.sample_emg_filtered[0])),
+#             len(ecg_gated_0) ,
+#         )
 
-    def test_gating_method_1(self):
-        ecg_gated_1 = gating(
-            self.sample_emg_filtered[0, :],
-            self.ecg_peaks,
-            gate_width=205,
-            method=1
-        )
+#     def test_gating_method_1(self):
+#         ecg_gated_1 = gating(
+#             self.sample_emg_filtered[0, :],
+#             self.ecg_peaks,
+#             gate_width=205,
+#             method=1
+#         )
 
-        self.assertEqual(
-            (len(self.sample_emg_filtered[0])),
-            len(ecg_gated_1) ,
-        )
+#         self.assertEqual(
+#             (len(self.sample_emg_filtered[0])),
+#             len(ecg_gated_1) ,
+#         )
     
-    def test_gating_method_2(self):
-        ecg_gated_2 = gating(self.sample_emg_filtered[0, :], self.ecg_peaks, gate_width=205, method=2)
+#     def test_gating_method_2(self):
+#         ecg_gated_2 = gating(self.sample_emg_filtered[0, :], self.ecg_peaks, gate_width=205, method=2)
 
-        self.assertEqual(
-            (len(self.sample_emg_filtered[0])),
-            len(ecg_gated_2) ,
-        )
+#         self.assertEqual(
+#             (len(self.sample_emg_filtered[0])),
+#             len(ecg_gated_2) ,
+#         )
 
-    def test_gating_method_2_no_prior_segment(self):
-        ecg_gated_2 = gating(self.sample_emg_filtered[0, :], [100], gate_width=205, method=2)
+#     def test_gating_method_2_no_prior_segment(self):
+#         ecg_gated_2 = gating(self.sample_emg_filtered[0, :], [100], gate_width=205, method=2)
 
-        self.assertFalse(
-            np.isnan(np.sum(ecg_gated_2))
-        )
+#         self.assertFalse(
+#             np.isnan(np.sum(ecg_gated_2))
+#         )
     
-    def test_gating_method_3(self):
-        ecg_peaks, _  = scipy.signal.find_peaks(self.sample_emg_filtered[0, :10*2048-1])
-        ecg_gated_3 = gating(self.sample_emg_filtered[0, :10*2048], ecg_peaks, gate_width=205, method=3)
+#     def test_gating_method_3(self):
+#         ecg_peaks, _  = scipy.signal.find_peaks(self.sample_emg_filtered[0, :10*2048-1])
+#         ecg_gated_3 = gating(self.sample_emg_filtered[0, :10*2048], ecg_peaks, gate_width=205, method=3)
 
-        self.assertEqual(
-            (len(self.sample_emg_filtered[0, :10*2048])),
-            len(ecg_gated_3) ,
-        )
+#         self.assertEqual(
+#             (len(self.sample_emg_filtered[0, :10*2048])),
+#             len(ecg_gated_3) ,
+#         )
 
 
 class TestMl(unittest.TestCase):
